+++ CHANGELOG.md
@@ -41,12 +41,18 @@ from `algosdk` and sends it to the network.
 - Added json_ref opcode to `Runtime`
 - Added support for foreign app account access in `Runtime`
 - Added guide for multisignature and signedTransaction usage in `Runtime`.
+- Added multisignatue verification in `Runtime`.
+- Added support for rekey account to multisignature in `Runtime`.
+- Added support to withdraw from account rekeyed to multisignature in `Runtime`.
+- Added new section in `Deployer` guide about helper methods across all deployers.
+- Added `SignTx`, `makeTx`, `makeAndSignTx`, `sendTxAndWait` helper methods in `Runtime` and `Deployer`.
 - Added support for teal v7 in `Runtime` and a test case.
 
 #### @algo-builder/web
 - Added support for logic signature to `executeTx` method of `Webmode` for AlgoSigner, MyAlgo Wallet and Wallet Connect.
 - Added `appendSignMultisigTransaction` function to `WebMode` for appending signature to multisig transaction in the algosigner.
 - Added `MultiSignature` support in `executeTx` method for `AlgoSigner`.
+- Added `SignTx`, `makeTx`, `makeAndSignTx`, `sendTxAndWait` helper methods in all webmodes.
 
 ### Bug Fixes